requirements.txt
fastapi
uvicorn
mysqlclient
python-multipart
python-dotenv
sqlalchemy
httpx
authlib
itsdangerous